placing viewsingleitem and editsingleitem into another component to make it cleaner
